+++ src/cloud/ops.rs
@@ -13,7 +13,7 @@ use crate::branding::BRANDING_CLOUD;
 use crate::cloud::client::{CloudClient, ErrorResponse};
 use crate::collect::Collector;
 use crate::options::CloudOptions;
-use crate::portable::status::{RemoteStatus, RemoteType};
+use crate::portable::instance::status::{RemoteStatus, RemoteType};
 use crate::question;
 
 const OPERATION_WAIT_TIME: Duration = Duration::from_secs(20 * 60);
