+++ src/portable/main.rs
@@ -1,49 +1,7 @@
 use crate::options::Options;
-use crate::portable::options::ServerInstanceCommand;
 use crate::portable::project::ProjectCommand;
 
-use crate::portable::backup;
-use crate::portable::control;
-use crate::portable::create;
-use crate::portable::credentials;
-use crate::portable::destroy;
-use crate::portable::link;
 use crate::portable::project;
-use crate::portable::reset_password;
-use crate::portable::resize;
-use crate::portable::revert;
-use crate::portable::status;
-use crate::portable::upgrade;
-use crate::portable::windows;
-
-pub fn instance_main(cmd: &ServerInstanceCommand, options: &Options) -> Result<(), anyhow::Error> {
-    use crate::portable::options::InstanceCommand::*;
-
-    match &cmd.subcommand {
-        Create(c) => create::create(c, options),
-        Destroy(c) => destroy::destroy(c, options),
-        ResetPassword(c) => reset_password::reset_password(c),
-        Link(c) => link::link(c, options),
-        List(c) if cfg!(windows) => windows::list(c, options),
-        List(c) => status::list(c, options),
-        Resize(c) => resize::resize(c, options),
-        Backup(c) => backup::backup(c, options),
-        Restore(c) => backup::restore(c, options),
-        ListBackups(c) => backup::list(c, options),
-        Upgrade(c) => upgrade::upgrade(c, options),
-        Start(c) => control::start(c),
-        Stop(c) => control::stop(c),
-        Restart(c) if cfg!(windows) => windows::restart(c),
-        Restart(c) => control::restart(c, options),
-        Logs(c) if cfg!(windows) => windows::logs(c),
-        Logs(c) => control::logs(c),
-        Revert(c) => revert::revert(c),
-        Unlink(c) => link::unlink(c),
-        Status(c) if cfg!(windows) => windows::status(c),
-        Status(c) => status::status(c, options),
-        Credentials(c) => credentials::show_credentials(options, c),
-    }
-}
 
 pub fn project_main(cmd: &ProjectCommand, options: &Options) -> anyhow::Result<()> {
     use crate::portable::project::Command::*;
