+++ src/portable/mod.rs
@@ -11,20 +11,10 @@ pub mod linux;
 pub mod macos;
 pub mod windows;
 
-mod backup;
-mod control;
-mod create;
-mod credentials;
-mod destroy;
 pub mod extension;
-mod link;
+pub mod instance;
 pub mod project;
-mod reset_password;
-mod resize;
-mod revert;
 pub mod server;
-pub mod status;
-mod upgrade;
 
-pub use main::{instance_main, project_main};
-pub use reset_password::password_hash;
+pub use instance::reset_password::password_hash;
+pub use main::project_main;
