+++ src/portable/windows.rs
@@ -24,15 +24,17 @@ use crate::commands::ExitCode;
 use crate::credentials;
 use crate::hint::HintExt;
 use crate::platform::{cache_dir, config_dir, tmp_file_path, wsl_dir};
-use crate::portable::control;
-use crate::portable::destroy;
 use crate::portable::exit_codes;
+use crate::portable::instance;
+use crate::portable::instance::control;
+use crate::portable::instance::create;
+use crate::portable::instance::destroy;
+use crate::portable::instance::status;
 use crate::portable::local::{write_json, InstanceInfo, NonLocalInstance, Paths};
-use crate::portable::options::{self, instance_arg, InstanceName, Logs, StartConf};
+use crate::portable::options::{self, instance_arg, InstanceName};
 use crate::portable::project;
 use crate::portable::repository::{self, download, PackageHash, PackageInfo};
 use crate::portable::server;
-use crate::portable::status::{self, Service};
 use crate::portable::ver;
 use crate::print::{self, msg, Highlight};
 use crate::process;
@@ -202,14 +204,14 @@ pub fn path_to_windows(path: &Path) -> anyhow::Result<PathBuf> {
 }
 
 pub fn create_instance(
-    options: &options::Create,
+    options: &create::Command,
     name: &str,
     port: u16,
     paths: &Paths,
 ) -> anyhow::Result<()> {
     let wsl = ensure_wsl()?;
 
-    let inner_options = options::Create {
+    let inner_options = create::Command {
         name: Some(InstanceName::Local(name.to_string())),
         port: Some(port),
         ..options.clone()
@@ -228,10 +230,10 @@ pub fn create_instance(
     Ok(())
 }
 
-pub fn destroy(options: &options::Destroy, name: &str) -> anyhow::Result<()> {
+pub fn destroy(options: &destroy::Command, name: &str) -> anyhow::Result<()> {
     let mut found = false;
     if let Some(wsl) = get_wsl()? {
-        let options = options::Destroy {
+        let options = destroy::Command {
             non_interactive: true,
             quiet: true,
             ..options.clone()
@@ -698,8 +700,8 @@ pub fn restart_service(_inst: &InstanceInfo) -> anyhow::Result<()> {
     anyhow::bail!("running as a service is not yet supported on Windows");
 }
 
-pub fn service_status(_inst: &str) -> Service {
-    Service::Inactive {
+pub fn service_status(_inst: &str) -> status::Service {
+    status::Service::Inactive {
         error: "running as a service is not yet supported on Windows".into(),
     }
 }
@@ -768,7 +770,7 @@ pub fn info(options: &server::Info) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn reset_password(options: &options::ResetPassword, name: &str) -> anyhow::Result<()> {
+pub fn reset_password(options: &instance::reset_password::Args, name: &str) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb()
             .arg("instance")
@@ -785,7 +787,7 @@ pub fn reset_password(options: &options::ResetPassword, name: &str) -> anyhow::R
     Ok(())
 }
 
-pub fn start(options: &options::Start, name: &str) -> anyhow::Result<()> {
+pub fn start(options: &control::Start, name: &str) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         if options.foreground {
             wsl.edgedb()
@@ -805,7 +807,7 @@ pub fn start(options: &options::Start, name: &str) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn stop(options: &options::Stop, name: &str) -> anyhow::Result<()> {
+pub fn stop(options: &control::Stop, name: &str) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         let service_file = service_file(name)?;
         fs::remove_file(&service_file)
@@ -825,7 +827,7 @@ pub fn stop(options: &options::Stop, name: &str) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn restart(options: &options::Restart) -> anyhow::Result<()> {
+pub fn restart(options: &control::Restart) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb()
             .arg("instance")
@@ -841,7 +843,7 @@ pub fn restart(options: &options::Restart) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn logs(options: &options::Logs) -> anyhow::Result<()> {
+pub fn logs(options: &control::Logs) -> anyhow::Result<()> {
     if let Some(wsl) = get_wsl()? {
         wsl.edgedb()
             .arg("instance")
@@ -857,7 +859,7 @@ pub fn logs(options: &options::Logs) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn status(options: &options::Status) -> anyhow::Result<()> {
+pub fn status(options: &status::Status) -> anyhow::Result<()> {
     if options.service {
         if let Some(wsl) = get_wsl()? {
             wsl.edgedb()
@@ -873,7 +875,7 @@ pub fn status(options: &options::Status) -> anyhow::Result<()> {
             return Err(ExitCode::new(exit_codes::INSTANCE_NOT_FOUND).into());
         }
     } else {
-        let inner_opts = options::Status {
+        let inner_opts = status::Status {
             quiet: true,
             ..options.clone()
         };
@@ -896,9 +898,9 @@ pub fn status(options: &options::Status) -> anyhow::Result<()> {
     Ok(())
 }
 
-fn list_local(options: &options::List) -> anyhow::Result<Vec<status::JsonStatus>> {
+fn list_local(options: &status::List) -> anyhow::Result<Vec<status::JsonStatus>> {
     if options.debug || options.extended {
-        let inner_opts = options::List {
+        let inner_opts = status::List {
             quiet: true,
             no_remote: true,
             ..options.clone()
@@ -911,7 +913,7 @@ fn list_local(options: &options::List) -> anyhow::Result<Vec<status::JsonStatus>
                 .run()?;
         }
     }
-    let inner_opts = options::List {
+    let inner_opts = status::List {
         no_remote: true,
         extended: false,
         debug: false,
@@ -933,7 +935,7 @@ fn list_local(options: &options::List) -> anyhow::Result<Vec<status::JsonStatus>
     Ok(local)
 }
 
-pub fn list(options: &options::List, opts: &crate::Options) -> anyhow::Result<()> {
+pub fn list(options: &status::List, opts: &crate::Options) -> anyhow::Result<()> {
     let errors = Collector::new();
     let local = match list_local(options) {
         Ok(local) => local,
@@ -1000,7 +1002,7 @@ pub fn list(options: &options::List, opts: &crate::Options) -> anyhow::Result<()
     }
 }
 
-pub fn upgrade(options: &options::Upgrade, name: &str) -> anyhow::Result<()> {
+pub fn upgrade(options: &instance::upgrade::Command, name: &str) -> anyhow::Result<()> {
     let wsl = try_get_wsl()?;
     wsl.edgedb()
         .arg("instance")
@@ -1012,7 +1014,7 @@ pub fn upgrade(options: &options::Upgrade, name: &str) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub fn revert(options: &options::Revert, name: &str) -> anyhow::Result<()> {
+pub fn revert(options: &instance::revert::Command, name: &str) -> anyhow::Result<()> {
     let wsl = try_get_wsl()?;
     wsl.edgedb()
         .arg("instance")
